+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to
 - libvirt: Introduce from ostools.
 - top-Makefile: Apply sub-dir to user targets.
 #### Added
+- shell: add align.sh, alias.sh.
 - cpp: Test `std::weak_ptr expired()`.
 - tkill(2): add tkill.c demo.
 - tgkill(2): add tgkill.c demo.